+++ liquid/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.6.4"
+__version__ = "0.7.0"
 
 from .mode import Mode
 from .filter import Filter
@@ -6,7 +6,7 @@ from .token import Token
 from .context import Context
 from .expression import Expression
 
-from .template import Template
 from .loaders import FileSystemLoader
 
 from .environment import Environment
+from .environment import Template